Try to smooth out animation with async task
There is a problem with FragmentStatePagerAdapter.
In my case I want to refresh the listview when user switch tab.
Problems associated with this are addressed here:
https://stackoverflow.com/questions/10849552/update-viewpager-dynamically/10852046#10852046
https://stackoverflow.com/questions/7263291/viewpager-pageradapter-not-updating-the-view
People suggest some solutions, like overriding the getItempostion or
instantiateItem. I chose the first approach.
It came with a caveat that all the view must be recreated.
Which results in a small lag when switching between tabs.
This approach is not efficient, yet I am unable to implement others.
Moving database query to background async task is just an attemp
to smooth animation.